pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "botgram"
version = "0.0.1"
authors = [
  { name="Manuele Pesenti", email="manuele@inventati.org" },
]
description = "Bot Telegram Package"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiogram == 3.1.1",
    "pydal",
    "psycopg2-binary",
    "python-dotenv"
]

# [project.urls]
# "Homepage" = "https://github.com/pypa/sampleproject"
# "Bug Tracker" = "https://github.com/pypa/sampleproject/issues"

[project.scripts]
run-botgram = "logbot.bot:run"